refactor: TimeValue 변환 로직을 utils/date로 분리
